Mend bad routes
Due to a state problem in the pretick - tick architecture, the current state of the cargo is already accounted for in newer routes made in the pretick, although the route still has to be altered in the tick.

This happens when consensus is being found for a new Parcel and at the same time a parcel was picked  up in the previous tick.